backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.1.1
opencv-python-headless==4.10.0.84
httpx==0.27.2
groq==0.11.0
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
